refactor: reorganize code to offload impl from dundle files
