+++ sapo/cli/__init__.py
@@ -1,3 +1,12 @@
 """Sapo CLI package."""
 
-__version__ = "0.1.0"
+from importlib.metadata import version
+
+try:
+    __version__ = version("sapo")
+except ImportError:
+    # Package is not installed
+    __version__ = "0.2.0"
+
+# Export the app for external use
+from .cli import app
